+++ plugins/3MFWriter/PCBDialog.qml
@@ -8,7 +8,6 @@ import QtQuick.Window 2.2
 
 import UM 1.5 as UM
 import Cura 1.1 as Cura
-import PCBWriter 1.0 as PCBWriter
 
 UM.Dialog
 {
