+++ plugins/3MFWriter/__init__.py
@@ -2,9 +2,12 @@
 # Uranium is released under the terms of the LGPLv3 or higher.
 import sys
 
+from PyQt6.QtQml import qmlRegisterType
+
 from UM.Logger import Logger
 try:
     from . import ThreeMFWriter
+    from .SettingsExportGroup import SettingsExportGroup
     threemf_writer_was_imported = True
 except ImportError:
     Logger.log("w", "Could not import ThreeMFWriter; libSavitar may be missing")
@@ -23,20 +26,36 @@ def getMetaData():
 
     if threemf_writer_was_imported:
         metaData["mesh_writer"] = {
-            "output": [{
-                "extension": "3mf",
-                "description": i18n_catalog.i18nc("@item:inlistbox", "3MF file"),
-                "mime_type": "application/vnd.ms-package.3dmanufacturing-3dmodel+xml",
-                "mode": ThreeMFWriter.ThreeMFWriter.OutputMode.BinaryMode
-            }]
+            "output": [
+                {
+                    "extension": "3mf",
+                    "description": i18n_catalog.i18nc("@item:inlistbox", "3MF file"),
+                    "mime_type": "application/vnd.ms-package.3dmanufacturing-3dmodel+xml",
+                    "mode": ThreeMFWriter.ThreeMFWriter.OutputMode.BinaryMode
+                },
+                {
+                    "extension": "pcb",
+                    "description": i18n_catalog.i18nc("@item:inlistbox", "PCB file"),
+                    "mime_type": "application/x-pcb",
+                    "mode": ThreeMFWriter.ThreeMFWriter.OutputMode.BinaryMode
+                }
+            ]
         }
         metaData["workspace_writer"] = {
-            "output": [{
-                "extension": workspace_extension,
-                "description": i18n_catalog.i18nc("@item:inlistbox", "Cura Project 3MF file"),
-                "mime_type": "application/vnd.ms-package.3dmanufacturing-3dmodel+xml",
-                "mode": ThreeMFWorkspaceWriter.ThreeMFWorkspaceWriter.OutputMode.BinaryMode
-            }]
+            "output": [
+                {
+                    "extension": workspace_extension,
+                    "description": i18n_catalog.i18nc("@item:inlistbox", "Cura Project 3MF file"),
+                    "mime_type": "application/vnd.ms-package.3dmanufacturing-3dmodel+xml",
+                    "mode": ThreeMFWorkspaceWriter.ThreeMFWorkspaceWriter.OutputMode.BinaryMode
+                },
+                {
+                    "extension": "pcb",
+                    "description": i18n_catalog.i18nc("@item:inlistbox", "Pre-Configured Batch file"),
+                    "mime_type": "application/x-pcb",
+                    "mode": ThreeMFWorkspaceWriter.ThreeMFWorkspaceWriter.OutputMode.BinaryMode
+                }
+            ]
         }
 
     return metaData
@@ -44,6 +63,8 @@ def getMetaData():
 
 def register(app):
     if "3MFWriter.ThreeMFWriter" in sys.modules:
+        qmlRegisterType(SettingsExportGroup, "ThreeMFWriter", 1, 0, "SettingsExportGroup")
+
         return {"mesh_writer": ThreeMFWriter.ThreeMFWriter(), 
                 "workspace_writer": ThreeMFWorkspaceWriter.ThreeMFWorkspaceWriter()}
     else:
