+++ resources/qml/Menus/FileMenu.qml
@@ -70,6 +70,18 @@ Cura.Menu
         enabled: UM.WorkspaceFileHandler.enabled
     }
 
+    Cura.MenuItem
+    {
+        id: savePCBMenu
+        text: catalog.i18nc("@title:menu menubar:file", "&Save PCB Project...")
+        enabled: UM.WorkspaceFileHandler.enabled
+        onTriggered:
+        {
+            var args = { "filter_by_machine": false, "file_type": "workspace", "preferred_mimetypes": "application/x-pcb" };
+            UM.OutputDeviceManager.requestWriteToDevice("local_file", PrintInformation.jobName, args)
+        }
+    }
+
     Cura.MenuSeparator { }
 
     UM.MeshWritersModel { id: meshWritersModel }
